requirements.txt
pandas == 2.2.2
joblib == 1.4.2
numpy == 1.26.4
sqlalchemy == 2.0.30
python-dotenv == 1.0.1
Flask-Cors== 4.0.1 
tensorflow-hub == 0.16.1
scikit-learn == 1.4.2
tensorflow == 2.16.1
tensorflow-hub == 0.16.1
Flask == 3.0.3